--- src/utils/anime/ChallengeInformation.ts
@@ -1,9 +1,0 @@
-export default interface ChallengeInformation {
-  name: string;
-  user: string;
-  animes: {
-    requerementId: number;
-    URL: string;
-    fields: string[];
-  }[];
-}
