+++ src/utils/anime/animeTypes.ts
@@ -28,6 +28,16 @@ export interface Challenge {
   requirements: Requirement[];
 }
 
+export interface ChallengeInformation {
+  name: string;
+  user: string;
+  animes: {
+    requerementId: number;
+    URL: string;
+    fields: string[];
+  }[];
+}
+
 export interface AnimeInformation {
   anime: string;
   user: string;
